==> code/15-chisq_rnatypes.Rmd <==
---
title: "chi square for RNA types"
author: "Megan Ewing"
date: "2025-01-30"
output: html_document
---

```{r}


# only need to run install once

# install.packages("dplyr")
# install.packages("tidyverse")
# install.packages("cowplot")
# install.packages("ggpubr")
# install.packages("rstatix")
# install.packages("BiocManager")
# BiocManager::install("DESeq2")
# install.packages("patchwork")

# packages -- may need to load each time project is open 

  library(dplyr)
  library(tidyverse)
  library(tidyr)
  library(data.table)
  library(ggplot2)
  library(RColorBrewer)
  library(cowplot)
  library(ggpubr)
  library(rstatix)
  library(DESeq2)

```

```{r}
#read in the count table from STAR (or DESEq2)<-ask Giles
#note: in BBEdit I replaced the "Type" information: replaced "protein-coding" with "mito" for the first 13 genes 
#which are the mitochondrial genes

counts<- read.csv("../data/chisq_rnatypes/count_data_mitoedit.csv", row.names = 1)
head(counts)

```

### Normalize Count Data Using DESeq2

preparation of count data

```{r}
# Read the star counts output file into R
starCounts <- read.csv("../output/STAR_count_data.csv", header=TRUE, row.names=1)
head(starCounts)

#subset from feature counts of just the count data
countData <- starCounts[, -(31:33)] 

# adjusting for sample 193
#confirm location
colnames(countData[16])

#we're moving it to first postion
countData <- countData[, c(16, 1:15, 17:ncol(countData))]
head(countData)

# specifying metadata
metaData <- data.frame(
  sample = c("M.C.193", "M.C.216", "M.C.218", "M.C.226", "M.C.306", "M.C.329", "M.C.334", "M.C.337", "M.C.339", "M.C.358", "M.C.360", "M.C.363", "M.C.373", "M.C.482","M.C.488", "M.T.18",  "M.T.20", "M.T.22", "M.T.235", "M.T.245", "M.T.29", "M.T.31", "M.T.32", "M.T.399", "M.T.43", "M.T.44", "M.T.500", "M.T.7", "M.T.83", "M.T.8"),
  treatment = c(rep("control", 15), rep("treatment", 15))
  )

# running deseq normalization precursor

# Create a DESeqDataSet object
dds <- DESeqDataSetFromMatrix(countData,
                              colData = metaData,
                              design = ~ treatment)

# prep for normalizing the data
dds <- DESeq(dds)
```

### normalization

```{r}
#normalizing the data using ratio of means (size factors)

dds <- estimateSizeFactors(dds)
sizeFactors(dds)
norm <- as.matrix(counts(dds, normalized=TRUE))
head(norm)

#saving as csv 
write.csv(norm, "../output/normalized_STAR.csv", row.names = T)

```

### need to bring back in gene names and rna type

```{r}

# importing norm as csv, because otherwise it stores as LHS or matrix (if specified). want as data frame for quicker processing
# plus, don't need to run the normalization / deseq steps each time if you just load normalized file
norm <- read.csv("../output/normalized_STAR.csv", row.names = 1)

# making dataframe of just genes and RNA type
types <- data.frame("gene" = rownames(counts), "Type" = counts$Type)
head(types)

# adding column to norm with gene names to prep for join
norm$gene <- rownames(norm)
head(norm)

# joining datframes to create new one with normalized count data and RNA type
norm_counts <- left_join(x = norm, y = types, by = "gene")
head(norm_counts)

# put genes back as row names
row.names(norm_counts) <- norm_counts$gene
head(norm_counts)

# fixing naming scheme for sample 193
colnames(norm_counts)[colnames(norm_counts) == 'M.T.193'] <- 'M.C.193'
head(norm_counts)
```

### Chi Square – Normalized Counts

```{r}

# create a summary table that sums the normalized counts of all control and all treatment samples 

summary_table <- norm_counts %>%
  group_by(Type) %>%
  summarise(
    control_sum = sum(M.C.193, M.C.216, M.C.218, M.C.226, M.C.306, M.C.329, M.C.334, M.C.337, M.C.339, M.C.358, M.C.360, M.C.363, M.C.373, M.C.482, M.C.488),
    treatment_sum = sum(M.T.18, M.T.20, M.T.22, M.T.235, M.T.245, M.T.31, M.T.32, M.T.399, M.T.43, M.T.44, M.T.500, M.T.7, M.T.8, M.T.83, M.T.29)
  )

head(summary_table)
```

### Convert to contingency table and run chi square

```{r}
# Convert to a matrix for statistical testing
contingency_table <- as.matrix(summary_table[, 2:3])
rownames(contingency_table) <- summary_table$Type

# Perform Chi-square test comparing control and treatment frequncies
chi_result <- chisq.test(contingency_table)
chi_result
```

### Visualizations

#### Chi Square Stacked Bar Chart

data prep

```{r}

# getting dataframe of proportions
props <- norm_counts  

# Apply function to columns with count data (1-30) that will divide each cell by the total number of reads for that sample (ie. cell # / sum of col)
props[, 1:30] <- lapply(norm_counts[, 1:30], function(col) col / sum(col))  

head(props)
```

```{r}

# summarizes the proportions by summing together all samples in each treatment
pie_data <- props %>%
  group_by(Type) %>%
  summarise(
    control_sum = sum(M.C.193, M.C.216, M.C.218, M.C.226, M.C.306, M.C.329, M.C.334, M.C.337, M.C.339, M.C.358, M.C.360, M.C.363, M.C.373, M.C.482, M.C.488),
    treatment_sum = sum(M.T.18, M.T.20, M.T.22, M.T.235, M.T.245, M.T.31, M.T.32, M.T.399, M.T.43, M.T.44, M.T.500, M.T.7, M.T.8, M.T.83, M.T.29)
  )

# divides the proportions by 15 (average proportion)
pie_data[, -1] <- pie_data[, -1] / 15  

head(pie_data)
```

chart itself

```{r}
library(ggplot2)
library(dplyr)
library(tidyr)
library(patchwork)

# Convert data to long format
pie_long <- pie_data %>%
  pivot_longer(cols = c(control_sum, treatment_sum), 
               names_to = "group", 
               values_to = "value")

# Create the stacked bar chart
stacked_bar <- ggplot(pie_long, aes(x = group, y = value, fill = factor(Type))) +
  geom_bar(stat = "identity", width = 0.5, position = "fill") +  # Stacked bar chart
  scale_y_continuous(labels = scales::percent) +  # Show percentage
  theme_minimal() +
  scale_x_discrete(labels=c("unprimed eggs", "primed eggs")) +
  labs(x = "", y = "Proportion of Reads", fill = "Type") +
  ggtitle("Proportion of Reads by Gene Type")

# Display plot
stacked_bar


```

#### Normalized Count Summary Table

```{r}
# install.packages("gt")
library("gt")
```

```{r}

# Create the table
gt_table <- gt(
  summary_table,
  rowname_col = "rowname",
  groupname_col = dplyr::group_vars(summary_table),
  process_md = FALSE,
  caption = NULL,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  auto_align = TRUE,
  id = NULL,
  locale = getOption("gt.locale"),
  row_group.sep = getOption("gt.row_group.sep", " - ")
) %>%
  cols_label(
    Type = "Type", 
    control_sum = "Unprimed",
    treatment_sum = "Primed"
  ) %>%
  fmt_number(
    columns = where(is.numeric),  # Format all numeric columns
    decimals = 2,  # Adjust decimal places as needed
    use_seps = TRUE  # Use thousands separators
  )

# Print the table
gt_table


```
